+++ host_database.cpp
@@ -15,9 +15,9 @@ host_database::host_database(long data_size, int dimension) {
     this->num_batches = 64;
     this->batch_size = this->data_size/this->num_batches;
 
-    host_database_B = (double**)malloc(num_batches * sizeof(double*)); //new double[data_size * dimension];
-    host_database_A = (double**)malloc(num_batches * sizeof(double*)); //new double[dimension];
-    host_database_prob_Y = (double**)malloc(num_batches * sizeof(double*)); //new double[dimension];
+    host_database_B = (float**)malloc(num_batches * sizeof(float*)); //new float[data_size * dimension];
+    host_database_A = (float**)malloc(num_batches * sizeof(float*)); //new float[dimension];
+    host_database_prob_Y = (float**)malloc(num_batches * sizeof(float*)); //new float[dimension];
 
 }
 
@@ -25,20 +25,20 @@ host_database::host_database(long data_size, int dimension) {
 void host_database::allocate() {
     /*allocating the memory for each matrix */
     for (int i=0;i<this->num_batches; i++){
-        host_database_B[i] = (double*)malloc(batch_size * dimension * sizeof(double)); //new double[data_size * dimension];
-        host_database_A[i] = (double*)malloc(batch_size * sizeof(double)); //new double[dimension];
-        host_database_prob_Y[i] = (double*)malloc(batch_size * sizeof(double)); //new double[dimension];
+        host_database_B[i] = (float*)malloc(batch_size * dimension * sizeof(float)); //new float[data_size * dimension];
+        host_database_A[i] = (float*)malloc(batch_size * sizeof(float)); //new float[dimension];
+        host_database_prob_Y[i] = (float*)malloc(batch_size * sizeof(float)); //new float[dimension];
     }
 
     // ---------------checking host memory  for error..............................
     if(host_database_B == nullptr)
-        mem_error("host_database_B", "vectmatmul", data_size * dimension, "double");
+        mem_error("host_database_B", "vectmatmul", data_size * dimension, "float");
 
     if(host_database_A == nullptr)
-        mem_error("host_database_A", "vectmatmul", data_size, "double");
+        mem_error("host_database_A", "vectmatmul", data_size, "float");
 
     if(host_database_prob_Y == nullptr)
-        mem_error("host_database_prob_Y", "vectmatmul", data_size, "double");
+        mem_error("host_database_prob_Y", "vectmatmul", data_size, "float");
 
 }
