+++ host_database.h
@@ -14,7 +14,7 @@ public:
     int num_batches; long batch_size;
 
 public:
-    double **host_database_B, **host_database_A, **host_database_prob_Y;
+    float **host_database_B, **host_database_A, **host_database_prob_Y;
 
 public:
     host_database(long, int);
