+++ utils.h
@@ -9,8 +9,8 @@
 /*mem error*/
 void mem_error(const std::string& arrayname, const std::string& benchmark, long len, const std::string& type);
 void print_on_screen(const std::string& program_name,double tsec,double gflops,long size,int flag);
-void fill_with_random_doubles(double* vec, long size);
-void fill_with_constant(double* vec, long size, double constant);
+void fill_with_random_doubles(float* vec, long size);
+void fill_with_constant(float* vec, long size, float constant);
 void relative_error(double* dRes, double* hRes, long size);
 double calculate_gflops(double &Tsec, long size);
